requirements.txt
awscli
loguru
mkdocs
pandas
pip
python-dotenv
requests
ruff
tqdm
typer
-e .
